backend-cop/requirements.txt
asgiref==3.8.1
attrs==24.3.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==2.0.12
click==8.1.8
Django==4.2.17
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==4.7.2
flake8==7.1.1
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==0.13.1
py==1.11.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2024.2
requests==2.26.0
sqlparse==0.5.3
toml==0.10.2
urllib3==1.26.20
psycopg2-binary
gunicorn
python-dotenv
drf-extra-fields>=3.4.0

